refactor: extract shared domain, random, and id helpers
