+++ internal/domain/types.go
@@ -1,0 +1,181 @@
+package domain
+
+// OpenAIRequest is the canonical Chat Completions request used by the proxy.
+type OpenAIRequest struct {
+	Model           string         `json:"model"`
+	Messages        []Message      `json:"messages"`
+	Stream          bool           `json:"stream"`
+	Temperature     *float64       `json:"temperature,omitempty"`
+	MaxTokens       *int           `json:"max_tokens,omitempty"`
+	TopP            *float64       `json:"top_p,omitempty"`
+	Thinking        any            `json:"thinking,omitempty"`
+	ReasoningEffort string         `json:"reasoning_effort,omitempty"`
+	ExtraBody       map[string]any `json:"extra_body,omitempty"`
+	Tools           []Tool         `json:"tools,omitempty"`
+	ToolChoice      any            `json:"tool_choice,omitempty"`
+}
+
+type Message struct {
+	Role             string     `json:"role,omitempty"`
+	Content          any        `json:"content,omitempty"`
+	ToolCalls        []ToolCall `json:"tool_calls,omitempty"`
+	ToolCallID       string     `json:"tool_call_id,omitempty"`
+	Name             string     `json:"name,omitempty"`
+	ReasoningContent *string    `json:"reasoning_content,omitempty"`
+}
+
+type ToolCall struct {
+	ID       string       `json:"id"`
+	Type     string       `json:"type"`
+	Function FunctionCall `json:"function"`
+}
+
+type FunctionCall struct {
+	Name      string `json:"name"`
+	Arguments string `json:"arguments"`
+}
+
+type Tool struct {
+	Type     string       `json:"type"`
+	Function ToolFunction `json:"function"`
+}
+
+type ToolFunction struct {
+	Name        string         `json:"name"`
+	Description string         `json:"description"`
+	Parameters  map[string]any `json:"parameters"`
+}
+
+type AppConfig struct {
+	ModelAlias           map[string]string `json:"model_alias"`
+	ReasoningEffortMap   map[string]string `json:"reasoning_effort_map"`
+	ForceDisableThinking bool              `json:"force_disable_thinking"`
+	MaxTokensCap         int               `json:"max_tokens_cap,omitempty"`
+	MaxTokensCapPerModel map[string]int    `json:"max_tokens_cap_per_model,omitempty"`
+	Socks5Proxies        []Socks5Proxy     `json:"socks5_proxies,omitempty"`
+	ActiveSocks5         string            `json:"active_socks5,omitempty"`
+	Socks5PaidDirect     bool              `json:"socks5_paid_direct,omitempty"`
+}
+
+type Socks5Proxy struct {
+	Addr     string `json:"addr"`
+	Username string `json:"username,omitempty"`
+	Password string `json:"password,omitempty"`
+	Name     string `json:"name,omitempty"`
+}
+
+type ClaudeRequest struct {
+	Model             string          `json:"model"`
+	Messages          []ClaudeMessage `json:"messages"`
+	System            any             `json:"system,omitempty"`
+	MaxTokens         *int            `json:"max_tokens,omitempty"`
+	Temperature       *float64        `json:"temperature,omitempty"`
+	TopP              *float64        `json:"top_p,omitempty"`
+	TopK              *int            `json:"top_k,omitempty"`
+	Stream            bool            `json:"stream,omitempty"`
+	Tools             []ClaudeTool    `json:"tools,omitempty"`
+	ToolChoice        any             `json:"tool_choice,omitempty"`
+	StopSequences     []string        `json:"stop_sequences,omitempty"`
+	Metadata          any             `json:"metadata,omitempty"`
+	Thinking          any             `json:"thinking,omitempty"`
+	OutputConfig      any             `json:"output_config,omitempty"`
+	ContextManagement any             `json:"context_management,omitempty"`
+}
+
+type ClaudeMessage struct {
+	Role    string `json:"role"`
+	Content any    `json:"content"`
+}
+
+type ClaudeContent struct {
+	Type      string `json:"type"`
+	Text      string `json:"text,omitempty"`
+	Thinking  string `json:"thinking,omitempty"`
+	Signature string `json:"signature,omitempty"`
+	Data      string `json:"data,omitempty"`
+	ID        string `json:"id,omitempty"`
+	Name      string `json:"name,omitempty"`
+	Input     any    `json:"input,omitempty"`
+	ToolUseID string `json:"tool_use_id,omitempty"`
+	Content   any    `json:"content,omitempty"`
+}
+
+type ClaudeTool struct {
+	Name         string `json:"name"`
+	Description  string `json:"description,omitempty"`
+	InputSchema  any    `json:"input_schema"`
+	Type         string `json:"type,omitempty"`
+	CacheControl any    `json:"cache_control,omitempty"`
+}
+
+type ClaudeResponse struct {
+	ID           string          `json:"id"`
+	Type         string          `json:"type"`
+	Role         string          `json:"role"`
+	Content      []ClaudeContent `json:"content"`
+	Model        string          `json:"model"`
+	StopReason   string          `json:"stop_reason"`
+	StopSequence *string         `json:"stop_sequence"`
+	StopDetails  any             `json:"stop_details,omitempty"`
+	Usage        ClaudeUsage     `json:"usage,omitempty"`
+}
+
+type ClaudeUsage map[string]any
+
+type ResponsesAPIRequest struct {
+	Model              string          `json:"model"`
+	Input              any             `json:"input"`
+	Messages           []Message       `json:"messages,omitempty"`
+	Instructions       string          `json:"instructions,omitempty"`
+	PreviousResponseID string          `json:"previous_response_id,omitempty"`
+	Stream             bool            `json:"stream,omitempty"`
+	Temperature        *float64        `json:"temperature,omitempty"`
+	MaxTokens          *int            `json:"max_output_tokens,omitempty"`
+	TopP               *float64        `json:"top_p,omitempty"`
+	FrequencyPenalty   *float64        `json:"frequency_penalty,omitempty"`
+	PresencePenalty    *float64        `json:"presence_penalty,omitempty"`
+	Reasoning          ReasonEffort    `json:"reasoning,omitempty"`
+	Include            []string        `json:"include,omitempty"`
+	Store              *bool           `json:"store,omitempty"`
+	Tools              []ResponsesTool `json:"tools,omitempty"`
+	ToolChoice         any             `json:"tool_choice,omitempty"`
+	ParallelToolCalls  *bool           `json:"parallel_tool_calls,omitempty"`
+	Stop               any             `json:"stop,omitempty"`
+	User               string          `json:"user,omitempty"`
+	StreamOptions      any             `json:"stream_options,omitempty"`
+	Metadata           any             `json:"metadata,omitempty"`
+	Text               any             `json:"text,omitempty"`
+	Truncation         string          `json:"truncation,omitempty"`
+	ServiceTier        string          `json:"service_tier,omitempty"`
+	PromptCacheKey     string          `json:"prompt_cache_key,omitempty"`
+	SafetyIdentifier   any             `json:"safety_identifier,omitempty"`
+	TopLogprobs        *int            `json:"top_logprobs,omitempty"`
+}
+
+type ResponsesTool struct {
+	Type            string         `json:"type"`
+	Name            string         `json:"name,omitempty"`
+	Description     string         `json:"description,omitempty"`
+	Parameters      map[string]any `json:"parameters,omitempty"`
+	Function        *ToolFunction  `json:"function,omitempty"`
+	ServerLabel     string         `json:"server_label,omitempty"`
+	ServerURL       string         `json:"server_url,omitempty"`
+	ConnectorID     string         `json:"connector_id,omitempty"`
+	Authorization   string         `json:"authorization,omitempty"`
+	AllowedTools    []string       `json:"allowed_tools,omitempty"`
+	RequireApproval any            `json:"require_approval,omitempty"`
+}
+
+type ReasonEffort struct {
+	Effort  string `json:"effort,omitempty"`
+	Summary string `json:"summary,omitempty"`
+	Mode    string `json:"mode,omitempty"`
+}
+
+type StoredResponseState struct {
+	Model        string          `json:"model"`
+	Instructions string          `json:"instructions,omitempty"`
+	Tools        []ResponsesTool `json:"tools,omitempty"`
+	ToolChoice   any             `json:"tool_choice,omitempty"`
+	Output       []any           `json:"output,omitempty"`
+}
