+++ internal/random/random.go
@@ -1,0 +1,25 @@
+package random
+
+import "crypto/rand"
+
+// String returns n random lowercase letters/digits.
+func String(n int) string {
+	const letters = "abcdefghijklmnopqrstuvwxyz0123456789"
+	b := make([]byte, n)
+	rand.Read(b)
+	for i := range b {
+		b[i] = letters[b[i]%byte(len(letters))]
+	}
+	return string(b)
+}
+
+// Hex returns n random hex characters.
+func Hex(n int) string {
+	const hex = "0123456789abcdef"
+	b := make([]byte, n)
+	rand.Read(b)
+	for i := range b {
+		b[i] = hex[b[i]%byte(len(hex))]
+	}
+	return string(b)
+}
